+++ src/uploader/fake.lisp
@@ -7,21 +7,21 @@
   (:import-from #:ultralisp/variables
                 #:get-dist-dir
                 #:get-clpi-dist-dir)
-  (:import-from #:metatilities
-                #:relative-pathname)
   (:import-from #:log)
   (:import-from #:ultralisp/utils
-                #:walk-dir))
+                #:walk-dir
+                #:relative-path))
 (in-package #:ultralisp/uploader/fake)
 
 
 (defmethod make-uploader ((type (eql :fake)) repo-type)
   (lambda (dir-or-file destination-path &key only-files)
-    (let* ((destination-path (relative-pathname (uiop:ensure-directory-pathname
-                                                 (ecase repo-type
-                                                   (:quicklisp (get-dist-dir))
-                                                   (:clpi (get-clpi-dist-dir))))
-                                                destination-path))
+    (let* ((destination-path (relative-path
+                              (uiop:ensure-directory-pathname
+                               (ecase repo-type
+                                 (:quicklisp (get-dist-dir))
+                                 (:clpi (get-clpi-dist-dir))))
+                              destination-path))
            (need-to-upload-p
              (make-files-inclusion-checker only-files)))
       (walk-dir (dir-or-file absolute relative)